feat: move engine to lib subpackage
