+++ libs/engine/src/lib.rs
@@ -1,0 +1,3 @@
+#![cfg_attr(not(test), no_std)]
+pub mod engine;
+pub mod motor;
