--- tests/src/lib.rs
@@ -1,11 +1,0 @@
-mod engine;
-mod motor;
-
-#[cfg(test)]
-mod tests {
-
-    #[test]
-    fn validate_test() {
-        assert_eq!(true, true);
-    }
-}
